preinstall: Return warnings as errors if RunChecks is already going to return an error
If RunChecks returns one or more errors, any warnings that have been
collected and which would normally be returned via the CheckResults
structure will be discarded. In the case where RunChecks is already
going to return one or more errors, it should also return any warnings
that occurred as errors as well.