preinstall: Add PPI actions to enable and clear the TPM
This adds 3 PPI actions for resolving various issues with the TPM:
- ActionEnableTPMViaFirmware, to enable the TPM.
- ActionEnableAndClearTPMViaFirmware, to enable (or ensure enabled) and
  clear the TPM.
- ActionClearTPMViaFirmware, to clear the TPM.

Once a PPI action has been submitted, the system needs to transition back
to the firmware for it to be handled. The `RunChecksContext.Run` API will
return an error kind that is either `ErrorKindReboot` or
`ErrorKindShutdown` to indicate how to transion back to firmware (known
as the state transition action). I've only ever seen "reboot" for this,
but the PPI can specify it to be "reboot", "shutdown" and some other
values that we ignore.

Once a PPI action has been submitted, calling `RunChecksContext.Run`
will eventually return the same error kind on subsequent calls, to tell
the caller reboot or shutdown. This is because the PPI operation may
resolve more than one error. It doesn't work like this right now because
go-tpm2's ppi.PPI interface does not provide a way to do it (see
https://github.com/canonical/secboot/issues/451.

The PPI indicates which PPI operations are supported, and whether those
operations require physical presence or not. If an operation is not
supported, then `RunChecksContext.Run` will not return the corresponding
Action. It is common for firmware implementations to only support one
of ActionEnableAndClearTPMViaFirmware and ActionClearTPMViaFirmware.