efi: Introduce WithAllowInsufficientDmaProtection API
WithAllowInsufficientDMAProtection can be supplied to AddPCRProfile to allow for
PCR7 including the "DMA Protection Disabled" event. While this reduces security,
it is required on some devices.

If this string is present in the event log, this option results in a creation of a
branched PCR profile that has two branches at the Firmware load stage one including
the event with the string, the another not.