efi/preinstall: Add a test that the lockout hierarchy is enabled
There is already a test that the dictionary attack protection hasn't
been triggered, which makes DA protected resources unavailable - this
results in a ErrTPMLockout error and a ErrorKindTPMDeviceLockout error
kind.

This adds another test that the actual lockout hierarchy is available,
which has its own dictionary attack protection. However, there is no
direct test for this other than to try authenticating the lockout
hierarchy, which we do by calling TPM2_DictionaryAttackLockReset if the
lockout hierarchy has no authorization value set. If the TPM returns a
TPM_RC_LOCKOUT warning, then we return a separate error for this -
ErrTPMLockoutLockedOut - and separate error kind -
ErrorKindTPMDeviceLockoutLockedOut, which indicate that the lockout
hierarchy is unavailable. This will remain unavailable until the
pre-programmed lockoutRecovery time expires or the TPM is cleared using
the platform hierarchy for authorization. There is no other means to
recover from this error.