efi/preinstall: Add explicit check that the current host system is EFI
If the host system is not EFI, the current tests will fail but that
error will be arbitarary depending on how the tests are ordered.

This adds a new explicit check that the system is EFI. If the host
system is not EFI, an ErrSystemNotEFI error will be returned, with the
error kind ErrorKindSystemNotEFI.

This also detects and wraps EFI variable access errors with
EFIVariableAccessError, which are all unexpected, considered fatal and
returned immediately, along with the error kind
ErrorKindEFIVariableAccess. This error kind is associated with an argument
- EFIVariableAccessErrorArg, which contains a description of the type of
access error, and maps to errors exported by github.com/canonical/go-efilib.