efi/preinstall: Provide a way to indicate that a required kernel module is missing
Some tests rely on kernel APIs that are currently shipped in modules as
opposed to being shipped in the main kernel image. The required modules
may not be loaded in all circumstances.

Whilst I don't think that secboot, as a dependency of snapd, should be directly
loading kernel modules, there needs to be a mechanism to tell snapd which
kernel modules are required in order for the tests to run successfully. The
required kernel modules are an implementation detail of secboot and may be
architecture / platform dependent as well, so snapd shouldn't have a list of
kernel modules to load. Instead, if a feature that requires a kernel module
cannot complete because it appears that the kernel module isn't loaded,
secboot will return a `MissingKernelModuleError` which describes the module
that snapd needs to subsequently load in order for the tests to complete.

As the need to load modules is an implementation detail of the preinstall
checks in general and shouldn't be exposed to the installer, there's no specific
error kind for this - it just gets converted to ErrorKindInternal and snapd is
expected to just resolve this without exposing it publicly.

This fixes https://github.com/canonical/secboot/issues/409