Add initial StorageContainer abstractions
This adds a new StorageContainerBackend interface to permit the core
secboot package to communicate with a registered package that handles
encrypted containers and keyslots (such as luks2) so that future code
in secboot will be agnostic to the type of storage container.

The StorageContainerBackend provides a way to probe a backend with a path
so that it can return a StorageContainer if appropriate. It also
provides a way to activate and deactivate an encrypted StorageContainer.

The StorageContainer provides a way to open a StorageContainerReader,
permitting code in secboot or outside of secboot to perform read only
operations on keyslots.

Eventually, the StorageContainer will provide a way to open a
StorageContainerWriter, permitting code in secboot or outside to
perform write operations to keyslots.

The backend implementation will be responsible for ensuring it is not
possible to have both StorageContainerReaders and StorageContainerWriters
open at the same time, and that whilst it will be possible to have an
aribtrary number of StorageContainerReaders open, there will only ever be
a single StorageContainerWriter that is permitted to be open.

The next steps after this PR are to:
- Fix the locking in internal/luks2 to not be "blocking" or
  "non-blocking". All current usage is blocking, so convert it to
  context based cancelation and expiration instead.
- Add a public luks2 package that registers a StorageContainerBackend
  implementation on import, and contains implementations of
  StorageContainer and StorageContainerReader. For now, these will just
  proxy to the existing APIs exported from secboot.
- Add an equivalent of InitializeLUKS2Container to the new luks2
  package, which will initially just proxy to the version currently
  exported from secboot.
- Add an initial implementation of the activation API with recovery key
  support (no PIN or passphrase support initially) which will make use
  of the newly implemented interfaces.
- Begin adding state to permit cross-checking of the primary keys during
  unlocking.
- Record unlocking state to the new activation API that can be serialized
  by snap-bootstrap.

Phase 2 tasks:
- Add a StorageContainerWriter interface for adding / deleting /
  renaming keyslots and updating keyslot metadata.
- Add an implementation of StorageContainerWriter to luks2, which for
  now will just proxy to the existing APIs exported from secboot.
- Update the AuthRequestor interface so that it is possible to request
  a passphrase, PIN and recovery key at the same time via a unified
  interface.
- Add PIN / passphrase support to the new activation API.

Phase 3 tasks (all post spike):
- Provide native implementations of the various StorageContainer
  interfaces in the luks2 package rather than proxying everything to the
  existing secboot API.
- Merge the internal/luksview package into the new luks2 package.
- Delete the LUKS2 specific parts of the core secboot package's API.
- Replace secboot.KeyDataReader with just an io.Reader. The concrete
  implementation on the backend side can still contain backend specific
  information, such as luks2 keyslot IDs.