internal/luks2: Remove LockMode in favour of context based cancelation and expiration
All existing users were using the blocking mode anyway, and I'll build
the ability to supply a context to new APIs.

Unfortunately, we do have to use flock in non-blocking mode now which means
that we have to retry it in a loop (with a random retry rate). It's not
possible to unblock a blocking flock call - I did try various things
including closing the underlying FD or using flock to unlock the same FD
from another goroutine, but neither of these worked. The retry delay is a
pseudo-random delay between 5ms and 100ms.