StorageContainers: Add some missing documentation.
This adds documentation to some types that had none, and also clarified
which interfaces are expected to be able to be used from multiple
goroutines and which interfaces are not.

It also adds some locking inside RegisterStorageContainerBackend.

I've left some notes discussing whether backends should be identified by
strings or whether they should instead be identified by arbitrary comparable
go types, in the same way that key/value pairs on context.Contextw work.