StorageContainers: update doc for NewStorageContainer
Rather than renaming the path argument to devicePath, I've clarified in
the documentation that not all storage containers may be represented by
block devices (eg, imagine we create a zfs storage backend in the future
that makes use of zfs native encryption and our own format for storing
and managing keyslots - in this case, the path probably isn't going to
be a device path. The same is true if we imagine that we might have a
fscrypt backend in the future.