keyring: Update keyring code to prepare for the new activation API
This makes several changes to the way that the keyring code works, in
preparation for the new activation API:
- It defines a better split between the code in keyring.go and the
  internal/keyring package. The internal/keyring package is now more of
  a shim around the kernel syscall interface.
- There is a new addKeyToUserKeyringLegacy function which replaces the
  deleted keyring.AddKeyToUserKeyring function. It's behaviour is
  unchanged and the existing activation code uses the replacement
  function.
- The StorageContainer interface has a new CredentialName method, to be
  used to associate a keyring key to a storage container. Whilst the
  luks2 implementation of this just returns the device path for both
  Path and CredentialName, I've added a new method to future proof the
  new code if we add additional storage backends in the future. Eg, if
  we add a fscrypt backend, then a path will not likely be consistent
  between early boot and runtime environments because they will be
  sensitive to mount points. Having a separate method would allow the
  backend to create an identifier that could be shared from early boot
  to runtime in this case, based on, eg, the device node and the path
  within the filesystem.
- There is a new addKeyToUserKeyring function which accepts a
  StorageContainer and makes use of the new CredentialName.
- There is a new single exported function for reading keys from the
  keyring, with a new KeyringKeyPurpose argument (GetKeyFromKernel). The
  new function is compatible with existing activation API, so that it
  is possible to use it from snapd with an older snap-bootstrap. The
  existing APIs and unit tests remain for now as long as the current API
  exists.

The new activation API will primarily make use of the new
addKeyToUserKeyring function, but it will support options to add
additional keys using the addKeyToUserKeyringLegacy function in the same
way that the existing activation API does. This means that it will be
possible to use the new API in snap-bootstrap when booting an older
version of snapd.

When the existing activation API is deprecated and removed, the
corresponding keyring related APIs will also be removed.

There is also a new internal/keyring/keyringtest package with some test
helpers for unit test code.

This also fixes some issues with the existing keyring unit tests which
meant they were being skipped in github. By doing this, some keyring
unit test support code (in internal/testutil) has been deleted. These
fixes are:
- Tests now consider that an invalidated key may have a temporary
  intermediate state due to how the kernel's garbage collector works
  (see the comment in keyringtest.InvalidateKeysAndWaitForGC).
- Some unit tests make use of the process and session keyrings, which
  are created on demand when required, but for the calling thread only.
  This means that the process keyring must be created in a
  single-thread execution context which isn't possible with go. The
  session keyring either needs to be set in the parent process or it
  must also be created in a single-thread execution context as well.
  The tests now get a process keyring and a new anonymous session keyring
  before the go runtime starts. Without this, the process or session
  keyring a goroutine accesses can be dependent on the OS thread it is
  running on.

The internal/keyring package has some protections to avoid triggering
creation and attachment of a new process keyring or a new anonymous
session keyring if there isn't one already. There are also some checks
to make sure that this doesn't happen inadvertently (eg, by something
using the syscall interface directly).