AuthRequestor: Merge the 2 methods into 1
The current interface has 2 methods - one to request a passphrase and
one to request a recovery key. In the future, we will also have PINs and
we want the ability for the user to supply any requested user credential
rather than having to wait to exhaust all of their passphrase/PIN
requests before they are able to enter a recovery key.

The new method can be asked to request any combination of passphrase,
PIN or recovery key using a flags argument.

One thing I'm not sure about yet is how best to make it possible to
localize messages for every combination of credential that can be
requested. My original idea was to use the text/template package which
gives the application a lot more flexiblity, but this does not work
because it brings in too many dependencies. For now, callers of
NewSystemdAuthRequestor supply a map of localized strings for every
combination of authentication types, but I'm not fully happy with this
and need to have a think about how to do something better.

Note that the argument names are different - sourceDevicePath is now
just "path" because the encrypted container may not always be a device.
The volumeName argument is just "name" and there will be an option for
the new activation API to customize this to be any human readable name.
The idea of "volumeName" only makes sense for backends that make use of
the device mapper to create new volumes, which of course, LUKS2 does.