tpm2: Introduce PCRPolicyDataError for PCR policy errors.
This makes it possible to distinguish between errors that can be
rectified by calling SealedKeyData.UpdatePCRProtectionPolicy, and errors
that require a keyslot to be replaced (which should be handled by
reprovisioning).

For v3 keys, PCR policy signature verification can fail as a result of
an incorrect approved digest, invalid signature or invalid PCR policy
reference. The approved digest and signature are updated when the PCR
policy is updated. This PR also ensures that the stored PCR policy
reference is refreshed whenever the PCR policy is updated. As a
consequence, the validity of the PCR policy reference is no longer
verified in keyData_v3.ValidateData.

Note that PCR policy signature verification could also fail because the
public key is invalid. This would still result in a PCRPolicyDataError
error being returned, but a subsequent attempt to update the PCR policy
would result in an InvalidKeyDataError being returned. In this case, the
keyslot would need to be replaced (likely by reprovisioning). A future
PR may improve this so that the stored public key is refreshed from the
supplied primary key.