luks2: Make the StorageContainer.Activate implementation work with any KeyslotInfo concrete type
The current implementation of StorageContainer.Activate returns an error
if the supplied KeyslotInfo isn't implemented by the concrete type from
this package (keyslotInfoImpl), as it uses this to obtain the LUKS2
keyslot ID.

This will be incompatible with the new activation API which will call
this with a KeyslotInfo of another type for externally loaded legacy key
data files. Instead, implement the existing behaviour of testing the key
against all of the LUKS2 keyslots if the supplied KeyslotInfo is not
implemented by the concrete type from this package.