preinstall: Accommodate poorly specified measurement order
PR #412 made it possible to use WithSecureBootPolicyProfile on devices
with pre-boot DMA protection disabled, which results in an additional
EV_EFI_ACTION event being measured to PCR7.

The problem is that this event is poorly specified - it's not in any TCG
specification or any other public specification. It is mentioned in some
tianocore documentation, but it doesn't define the measurement order and
there is no reference implementation in EDK2.

We've seen 2 logs with this event being measured in different places. In
addition to this, the event uses a NULL terminated string in one of
these when the TCG PC Client PFP spec says that these strings aren't
NULL terminated (this is already accomodated by the current
implementation though).

In order to work reliably across different firmware implementations,
this needs to accommodate the different measurement orders that could be
expected.

Fixes: #440.