efi: Permit pre-OS image verification events to be measured before the separator
The measurements to PCR7 are divided between secure boot configuration
(EV_EFI_VARIABLE_DRIVER_CONFIG) and secure boot verification
(EV_EFI_VARIABLE_AUTHORITY), separated by a single EV_SEPARATOR event.
This is true for all firmware implementations when ignoring the
existence of UEFI drivers that can loaded from addon cards, and system
preparation applications, both of which are verified and executed as
part of the pre-OS environment.

However, the TCG Client PFP spec is a little bit vague with regards to
how verification of third party pre-OS components is meant to work.

EDK2 measures the EV_SEPARATOR in PCR7 as soon as it has measured the
secure boot configuration, as opposed to waiting until the transition to
OS-present to do this. As a result, EV_EFI_VARIABLE_AUTHORITY events
associated with pre-OS components are measured *after* the EV_SEPARATOR
with EDK2 (but still in the pre-OS environment). In EDK2, the
EV_SEPARATOR in PCR7 is the boundary between the secure boot configuration,
and the verification of third party code. Comments in the EDK2 source
code indicate that the EV_SEPARATOR event in PCR7 is meant to indicate
the boundary between secure boot configuration and image verification.

Some other firmware implementations measure the EV_SEPARATOR in PCR7 as
part of the transition to OS-present, as opposed to doing it immediately
after the secure boot configuration is measured. As a result,
EV_EFI_VARIABLE_AUTHORITY events associated with pre-OS components are
measured *before* the EV_SEPARATOR with these firmware implementations.
With these firmware implementations, the EV_SEPARATOR in PCR7 is the
boundary between the pre-OS environment and the OS-present environment,
as it is with PCRs 0-6. EV_EFI_VARIABLE_AUTHORITY events associated with
OS components are still measured after the separator.

We currently only support platforms that behave like EDK2.

As the TCG spec is vague about this, we should support both measurement
sequences. Note that there is only a difference between measurement
sequences on systems that load third party code as part of the pre-OS
environment (ie, UEFI drivers loaded from addon cards / system
preparation applications).

This partially fixes https://github.com/canonical/secboot/issues/410
(corresponding changes in efi/preinstall are also required)