preinstall: Rename MissingKernelModuleError String method
A fmt.Stringer implementation might not necessarily return the
piece of information we want (eg, imagine it saying 'missing kernel
module: "msr"'), so rename the method to "Module" instead, documented
as returning only the module name string. We don't need a fmt.Stringer
implementation for this anyway, as it's already an error.