multiple: add support for FIDO2 related actions in platform agnostic API
platforms that implement the FIDO2ProviderInterface will need to
implement the GetSymmetricKey function which will be used by the FIDO2
platform in order to retrieve a symmetric key. This symmetric key will
then be used in conjuction with a FIDO2 authenticator to derive the
final symmetric key which will be used for unlocking the disk unlocking
key.

Platform agnostic API's Keydata was extended with a new function
RecoverFIDO2Salt() which will call the handler's GetSymmetricKey if the
KeyData indicates that its handler is a platform which requires FIDO2.