fido2/tpm2: add the new tpm2-fido2 platform
the tpm2-fido2 platform extends the existing plain tpm2 platform to add
support for FIDO2 authenticators. It is essentially implemented as a
shim platform which first unseals the necessary key material from the
TPM and then passes it on to the FIDO2 platform in order to get the
final key material required for unlocking the disk unlocking key.