preinstall: Add ActionProceed action
There are some error kinds that don't necessarily have to be resolved
and don't have to prevent FDE from being enabled, but can be used to
provide information to a user so that they can evaluate them and
determine whether they would like to proceed at their own risk. Eg, if
the system is executing third party code before the OS and the user
cannot or does not want to disable it, then there is an increased risk
of PCR fragility as a result of those third party components being
updated by something other than the OS. If an error indicates that
pre-boot DMA remapping is disabled and the user cannot enable it, then
there is an increased risk of a locally present adversary obtaining
secrets or altering code execution via external ports.

This introduces a new action - ActionProceed, which can be executed only
if all of the previously returned error kinds permit this action. When
used, a CheckFlags flag will be added for each unresolved error kind,
which has the effect of ingoring those errors and making them warnings
instead.