efi/preinstall: Add actions to clear the TPM using TPM2_Clear
This adds 2 actions that can be used to clear a TPM using the TPM2_Clear
command, which can be used if owner clear has not been disabled and
doesn't require a reboot. The 2 actions are:
- ActionClearTPMSimple which accepts no arguments and is only available
  if the lockout hierarchy has an empty authorization value. This one is
  suitable for exposing via the desktop installer UI.
- ActionClearTPM which can accept an authorization value, but is not
  really suitable for the desktop installer UI as that shouldn't expose
  the concept of authorization values. It may be suitable for a more
  advanced commandline tool or server systems.

They are separate actions so that ActionClearTPMSimple can be marked
separately as unavailable if the lockout hierarchy has a non-empty
authorization value.

I did want to add actions to clear hierarchy ownership, however,
clearing authorization values isn't suitable for the desktop installer
UI, and whilst clearing policies may be suitable in the case where a
hierarchy has an empty authorization value, the way this error is
returned as a combined error with information about all hierarchies
doesn't map well to the way that actions work. I might address this in a
future PR, although it might be better to just recommend clearing the
TPM in this case anyway.