efi/preinstall: Fix a panic that can happen when secure boot is disabled
If secure boot is disabled and the SbatLevel UEFI variable didn't exist
when shim executed, there will be no EV_EFI_VARIABLE_AUTHORITY events in
PCR7. At least one of these is required by
checkSecureBootPolicyPCRForDegradedSettings in order for the function to
return rather than the loop terminating and then hitting a panic at the
end of the function.