Support unlocking with ActivateContext using v1 platform keys
The ActivateContext API assumes that platform keys are all v2 keys. For
backwards compatibility, we need to support v1 platform keys. The
properties of v1 platform keys differ in the following ways:
- They are associated with one or more authorized snap models, so they
  need to be checked against the current boot model.
- Their primary keys cannot be crosschecked to demonstrate that 2 storage
  containers are bound to each other. In order to support v1 platform
  keys, there is a new WillCheckStorageContainerBinding option which can
  be used when the caller will use another mechanism to demonstrage that
  2 storage containers are bound to each other.

The WillCheckStorageContainerBinding disables the primary key
crosscheck. It should only be used when it is known that a storage
container will be unlocked using a v1 platform key. When used, the caller
must take steps to verify that the storage container being unlocked is
bound to those that have already been unlocked.

If one or more storage containers are unlocked with v1 platform keys,
then subsequent storage containers must be unlocked by a key that is
protected with a platform registered with the
PlatformProtectedByStorageContainer flag, unless the
WillCheckStorageContainerBinding option is supplied.

If one or more storage container are unlocked with new platform keys,
then subsequent storage containers must also be unlocked with new
platform keys unless the WillCheckStorageContainerBinding option is
supplied.