Add WithExternalUnlockKey option
This adds a new WithExternalUnlockKey option which provides a way to
supply a plain unlock key that can be used to attempt to unlock a
storage container.

External unlock keys don't have a primary key. In order to demonstrate
that a storage container is bound to those that have already been
unlocked, the caller must declare that the unlock key was recovered from
an already unlocked storage container, using the
ExternalUnlockKeyFromStorageContainer enum.