efi/preinstall: only indicate the mei_me module is missing if the ME PCI device exists
A MissingKernelModuleError error with the value "mei_me" should only be
returned if the ME device cannot be found but the PCI device exists. If
the PCI device does not exist, then UnsupportedPlatformError is a more
appropriate error to return.

This also ensures that in the case where more than one device exists in
the mei subsystem, we pick the one associated with the ME (a PCI device
that is bound to the mei_me module). On some platforms, there may be
more than one device in the mei subsystem, eg, Intel's Vision Sensing
Controller, which is a platform device that is bound to the intel_vsc
module.

Note that the code in internal/efi has become a bit of a dumping ground
for things that aren't related to EFI, and I'm going to split this up in
a future PR.