efi/preinstall: Add tests for AMD Platform Secure Boot
This adds a check that Platform Secure Boot is enabled on AMD platforms,
by reading an attribute from the PSP that is exposed by the kernel in
sysfs, and which indicates whether the OTP fuses are blown - after which
PSB is enabled. Once enabled, the PSP firmware verifies the initial boot
block of the firmware that executes on the main AMD CPU.

What isn't clear at the moment is whether the PSP firmware or UEFI
firmware is the root-of-trust for measurement. If it is the PSP
firmware, then a future update can add support for platforms without
PSB enabled, as we would still have a valid root-of-trust for
measurement and we would seal keys against PCR0 in this case.