efi/preinstall: Provide information about addon drivers and sysprep apps
This adds arguments for errors that indicate the presence of any addon
drivers or system preparation applications. This includes the loaded
image format (addon drivers can be firmware blobs or EFI images), a
description from the associated load option if there is one (except for
addon drivers loaded from an option ROM or a firmware blob), the EFI
device path (except for firmware blobs) and the image digest.

This also replaces the "value-added-retailer supplied driver"
terminology for "addon driver" instead, renaming the associated errors
and error kind accordingly, as I think this makes more sense.

Fixes: FR-11911
Fixes: https://github.com/canonical/secboot/issues/450