efi/preinstall: Support TPMs where TPM_PT_PS_FAMILY_INDICATOR is not TPM_PS_PC
We check that the TPM is a PC-Client device because we can make
assumptions about its feature set based on what is in the TCG PC-Client
Platform TPM Profile spec. Some TPMs don't return the correct value
though - eg, some return TPM_PS_PDA. There is already a workaround for
the swtpm case which also returns an unexpected value.

This replaces the existing workaround and adds explicit feature
detection for these TPMs.

Fixes: FR-11806
Fixes: https://github.com/canonical/secboot/issues/448
Fixes: https://github.com/canonical/secboot/issues/460