efi/preinstall: Record accepted errors in CheckResult
This records errors accepted via ActionProceed to a new AcceptedErrors
field of CheckResult.

There are 2 uses for this:
- To use accepted errors to turn on options when generating PCR profiles.
  The WithAllowInsufficientDmaProtection() opt-in makes use of this new
  field with this PR. Future PRs will add additional opt-ins which make
  use of this new field.
- To compare future check results against, in order to determine whether
  anything has changed.

At the moment, only the accepted error kinds are retained. However, the
new field is a map of ErrorKind to arbitrary JSON value so that a future
update can use the arbitrary JSON value to record arguments that can
limit the scope of an accepted error.

Fixes: FR-11919